backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
anthropic>=0.34.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
langgraph>=0.2.0
langchain-core>=0.3.0
pdfplumber>=0.11.0
python-docx>=1.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
